redo sep

+++ analysis_spi/sep_grid_SPI_compute.py
@@ -104,21 +104,22 @@ constructed_time_axis = np.append(np.arange(-132*100,0), np.arange(132*100))
 for iy,y in enumerate(lat):
 	for ix,x in enumerate(lon):
 		if land_mask[y,x] != 1 and y>=0:
-			start = time.time()
 			print(y,x)
 			grid_file_name = working_path+'grid_level/'+str(y)+'_'+str(x)+'.txt'
-			tmp = np.append(big_merge_hist[:,iy,ix],big_merge_fut[:,iy,ix])
-			csv = open(grid_file_name,'w')
-			csv.write(';'.join([str(i) for i in tmp]))
-			csv.close()
+			if os.path.file(grid_file_name.replace('.txt','_SPI3.txt')) == False:
+				start = time.time()
+				tmp = np.append(big_merge_hist[:,iy,ix],big_merge_fut[:,iy,ix])
+				csv = open(grid_file_name,'w')
+				csv.write(';'.join([str(i) for i in tmp]))
+				csv.close()
 
-			print(time.time() -start)
+				print(time.time() -start)
 
-			result=try_several_times('Rscript /global/homes/p/pepflei/persistence_in_models/analysis_add/add_61_SPI_single.r '+\
-				grid_file_name+' '+\
-				grid_file_name.replace('.txt','_SPI3.txt'),1,1000)
+				result=try_several_times('Rscript /global/homes/p/pepflei/persistence_in_models/analysis_add/add_61_SPI_single.r '+\
+					grid_file_name+' '+\
+					grid_file_name.replace('.txt','_SPI3.txt'),1,4000)
 
-			print(time.time() -start)
+				print(time.time() -start)
 
 
 '''
